Refactored ws auth code
