+++ frontend/src/chat/interfaces/chat.ts
@@ -10,12 +10,6 @@ export type MessageHandler = (msg: string, name: string) => void
 
 export type EmitHandler = (error: any, response: any) => void
 
-// export interface ChatSettings {
-// 	roomId: string
-// 	name: string
-// 	messageHandler: MessageHandler
-// }
-
 export interface ChatController {
 	sendMessage: (msg: string) => void
 }
