+++ frontend/src/config.ts
@@ -4,6 +4,4 @@ export const CLIENT_ID = process.env.CLIENT_ID
 // @ts-ignore
 export const CHAT_HOST = process.env.CHAT_HOST
 
-export const CHAT_HOST_FRONTEND = 'http://localhost'
-
 export const CHAT_OPTIONS = {transports: ['websocket']}
