+++ frontend/src/init/main.ts
@@ -1,8 +1,9 @@
-import type { AuthSettings } from '../chat/interfaces/chat'
+
 import {CLIENT_ID} from '../config'
 import appIcon from './icon'
 
-export let authSettings: AuthSettings| undefined
+import type { AuthSettings } from '../chat/interfaces/chat'
+import { getAuthSettings } from './auth'
 
 const initChat = (breakoutChatRoomId: string) => {
 	miro.__setRuntimeState({
@@ -46,57 +47,32 @@ const handleAddChatClick = async () => {
 }
 
 const initPlugin = async () => {
-	authSettings = await getAuthSettings()
-	if (!authSettings) return
-	// @ts-ignore
-	miro.addListener(miro.enums.event.SELECTION_UPDATED, async () => {
-		const widgets = await miro.board.selection.get()
-		if (widgets.length === 1 && widgets[0].metadata[CLIENT_ID]?.isBreakoutChatButton) {
-			initChat(widgets[0].id)
-		}
-	})
+	const authSettings: AuthSettings|undefined = await getAuthSettings()
+	let onClick: () => Promise<void>
+
+	if(authSettings) {
+		// @ts-ignore
+		miro.addListener(miro.enums.event.SELECTION_UPDATED, async () => {
+			const widgets = await miro.board.selection.get()
+			if (widgets.length === 1 && widgets[0].metadata[CLIENT_ID]?.isBreakoutChatButton) {
+				initChat(widgets[0].id)
+			}
+		})
+		onClick = handleAddChatClick
+	} else
+		onClick = () => miro.showNotification(
+			'We\'re sorry, but your permissions for this board are insufficient to use the chat functionality'
+		)
 
 	await miro.initialize({
 		extensionPoints: {
 			bottomBar: {
 				title: 'Create a new breakout chat',
 				svgIcon: appIcon,
-				onClick: handleAddChatClick,
+				onClick
 			},
 		},
 	})
-}
-
-const getCurrentUserName = async () => {
-	const id = await miro.currentUser.getId()
-	// @ts-ignore
-	const onlineUsers = await miro.board.getOnlineUsers()
-
-	return onlineUsers.find((user) => user.id === id)?.name
-}
-
-const getAuthSettings = async (): Promise<AuthSettings | undefined> => {
-	const username = await getCurrentUserName()
-	const token = await miro.getToken()
-	const boardInfo = await miro.board.info.get()
-	const currentBoardPermissions = boardInfo.currentUserPermissions
-
-	// Somewhat debatable mapping from 'currentBoardPermissions' to 'user may
-	// use the chat functionality': I've decided that at least one of the
-	// following three board permissions needs to be set: "EDIT_INFO",
-	// "EDIT_CONTENT", "EDIT_COMMENTS".
-	// NB! Decide with the rest of the team whether this is strict enough.
-	const userMayEdit = currentBoardPermissions.indexOf('EDIT_INFO') > - 1
-		|| currentBoardPermissions.indexOf('EDIT_CONTENT') > - 1
-		|| currentBoardPermissions.indexOf('EDIT_COMMENTS') > - 1
-
-	if (username && token && boardInfo.id && userMayEdit)
-		return {
-			username,
-			token,
-			boardId: boardInfo.id,
-			mayReadWriteBoard: true
-		}
 }
 
 miro.onReady(async () => {
